fix(CreatureHelper): Correct integer overflow in CalculateTicksForClientTicks
The `CalculateTicksForClientTicks` method could overflow when a large `clientTicks` value was provided, causing an incorrect calculation. This commit fixes the issue by casting the `clientTicks` value to a `long` before the multiplication, preventing the overflow.

A new test case has been added to `CreatureHelperTests.cs` to verify the fix and prevent future regressions.

+++ Hagalaz.Game.Utilities/CreatureHelper.cs
@@ -32,6 +32,6 @@ namespace Hagalaz.Game.Utilities
         /// </summary>
         /// <param name="clientTicks">The client ticks.</param>
         /// <returns>System.Int32.</returns>
-        public static int CalculateTicksForClientTicks(int clientTicks) => (clientTicks * 20 + 599) / 600;
+        public static int CalculateTicksForClientTicks(int clientTicks) => (int)((clientTicks * 20L + 599) / 600);
     }
 }
